server/requirements.txt
absl-py
aniso8601==8.0.0
autopep8
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1